# docs/contracts/knowledge-v0.openapi.yaml
openapi: 3.0.3
info:
  title: TenderHack knowledge service — contract v0
  version: "0.1.0"
  description: >
    Frozen contract between the .NET support core (`api`/`api-worker`) and the Python
    knowledge & inference service (`knowledge`/`knowledge-worker`). See
    docs/contracts/knowledge-v0.md for normative rules (headers, error taxonomy,
    idempotency, versioning). This file is the source of truth for NSwag client
    generation until `knowledge` emits its own OpenAPI from FastAPI; at that point the
    two must stay in lockstep or the change is a contract-breaking change (see §6 of
    the companion markdown).
servers:
  - url: http://knowledge:8000
paths:
...
  /v0/understand:
    post:
      operationId: understand
      summary: Normalize raw user text and extract entities/exact codes.
      parameters:
        - $ref: "#/components/parameters/XTraceId"
        - $ref: "#/components/parameters/XCaseId"
        - $ref: "#/components/parameters/XTurnId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UnderstandRequest"
      responses:
        "200":
          description: Understanding result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UnderstandResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ModelUnavailable"
  /v0/moderation/context:
    post:
      operationId: moderationContext
      summary: LLM ambiguity check for a deterministic rule match. Never returns a moderation outcome.
      parameters:
        - $ref: "#/components/parameters/XTraceId"
        - $ref: "#/components/parameters/XCaseId"
        - $ref: "#/components/parameters/XTurnId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModerationContextRequest"
      responses:
        "200":
          description: Ambiguity assessment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModerationContextResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ModelUnavailable"
  /v0/retrieve:
    post:
      operationId: retrieve
      summary: Hybrid retrieval (exact + FTS + dense + RRF + rerank) against one corpus.
      parameters:
        - $ref: "#/components/parameters/XTraceId"
        - $ref: "#/components/parameters/XCaseId"
        - $ref: "#/components/parameters/XTurnId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RetrieveRequest"
      responses:
        "200":
          description: Ranked candidate fragments.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RetrieveResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/UnknownSnapshot"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ModelUnavailable"
  /v0/answerability:
    post:
      operationId: answerability
      summary: Assess whether retrieved evidence is sufficient to answer, without producing a decision.
      parameters:
        - $ref: "#/components/parameters/XTraceId"
        - $ref: "#/components/parameters/XCaseId"
        - $ref: "#/components/parameters/XTurnId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnswerabilityRequest"
      responses:
        "200":
          description: Evidence sufficiency assessment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnswerabilityResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/UnknownSnapshot"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ModelUnavailable"
  /v0/draft:
    post:
      operationId: draft
      summary: Generate a structured draft answer with claims tied to evidence fragments.
      parameters:
        - $ref: "#/components/parameters/XTraceId"
        - $ref: "#/components/parameters/XCaseId"
        - $ref: "#/components/parameters/XTurnId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DraftRequest"
      responses:
        "200":
          description: Draft with claims.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DraftResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/UnknownSnapshot"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/ModelError"
        "503":
          $ref: "#/components/responses/ModelUnavailable"
  /v0/verify:
    post:
      operationId: verify
      summary: Verify each claim of a draft against its cited evidence fragments.
      parameters:
        - $ref: "#/components/parameters/XTraceId"
        - $ref: "#/components/parameters/XCaseId"
        - $ref: "#/components/parameters/XTurnId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VerifyRequest"
      responses:
        "200":
          description: Per-claim verification result.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VerifyResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/UnknownSnapshot"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
        "503":
          $ref: "#/components/responses/ModelUnavailable"
  /v0/sources/{fragment_id}:
    get:
      operationId: getSource
      summary: Resolve a fragment to its original document/page/anchor text for the source drawer.
      parameters:
        - $ref: "#/components/parameters/XTraceId"
        - name: fragment_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Source fragment.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceResponse"
        "404":
          $ref: "#/components/responses/UnknownFragment"
        "500":
          $ref: "#/components/responses/InternalError"
  /v0/materials:
    get:
      operationId: listMaterials
      summary: List documents in the current normative snapshot (Materials tab, E3).
      parameters:
        - $ref: "#/components/parameters/XTraceId"
      responses:
        "200":
          description: Materials in the current normative snapshot.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaterialsResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /v0/materials/{document_id}/sections:
    get:
      operationId: listMaterialSections
      summary: Table of contents for one document in the current normative snapshot (E3).
      parameters:
        - $ref: "#/components/parameters/XTraceId"
        - name: document_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Sections in page order; a fragment with no section sorts last.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MaterialSectionsResponse"
        "404":
          $ref: "#/components/responses/UnknownDocument"
        "500":
          $ref: "#/components/responses/InternalError"
  /v0/snapshots/current:
    get:
      operationId: getCurrentSnapshot
      summary: Current active knowledge base snapshot.
      parameters:
        - $ref: "#/components/parameters/XTraceId"
      responses:
        "200":
          description: Current snapshot metadata.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SnapshotResponse"
        "500":
          $ref: "#/components/responses/InternalError"
  /v0/quality/turns:
    post:
      operationId: pushQualityTurn
      summary: >
        Push the full fact set of one completed turn (sent by api-worker via outbox,
        at-least-once, idempotent by turn_id+revision).
      parameters:
        - $ref: "#/components/parameters/XTraceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QualityTurnPush"
      responses:
        "202":
          description: Stored (not yet evaluated).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcceptedResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /v0/quality/feedback:
    post:
      operationId: pushQualityFeedback
      summary: >
        Push one feedback event (sent by api-worker via outbox, at-least-once,
        idempotent by feedback_id).
      parameters:
        - $ref: "#/components/parameters/XTraceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QualityFeedbackPush"
      responses:
        "202":
          description: Stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcceptedResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
  /v0/quality/completions:
    post:
      operationId: pushQualityCompletion
      summary: >
        Push the completion fact of one case (sent by api-worker via outbox,
        at-least-once, idempotent by case_id). Additive v0 endpoint (2026-09-12).
      parameters:
        - $ref: "#/components/parameters/XTraceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QualityCompletionPush"
      responses:
        "202":
          description: Stored.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AcceptedResponse"
        "422":
          $ref: "#/components/responses/ValidationError"
        "500":
          $ref: "#/components/responses/InternalError"
...
  /v0/quality/issue-groups:
    get:
      operationId: getIssueGroups
      summary: Read-only repeated-problem groups (proxied by api with server-side authz).
      parameters:
        - $ref: "#/components/parameters/XTraceId"
      responses:
        "200":
          description: Issue groups.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/IssueGroupsResponse"
        "500":
          $ref: "#/components/responses/InternalError"
components:
  parameters:
    XTraceId:
      name: X-Trace-Id
      in: header
      required: true
      schema:
        type: string
        format: uuid
    XCaseId:
      name: X-Case-Id
      in: header
      required: true
      schema:
        type: string
    XTurnId:
      name: X-Turn-Id
      in: header
      required: true
      schema:
        type: string
...
  schemas:
...
    UnderstandRequest:
      type: object
      required: [text]
      properties:
        text:
          type: string
        prior_turn_summary:
          type: string
          nullable: true
    UnderstandResponse:
      type: object
      required: [normalized_text, entities, exact_codes, language_flags]
      properties:
        normalized_text:
          type: string
        entities:
          type: array
          items:
            $ref: "#/components/schemas/Entity"
        exact_codes:
          type: array
          items:
            type: string
        language_flags:
          $ref: "#/components/schemas/LanguageFlags"
...